pyproject.toml
[tool.poetry]
name = "pystarport"
version = "0.2.5"
description = "Spawn local devnets for cosmos-sdk chains"
authors = ["chain-dev-team <chain-dev@crypto.com>"]
license = "Apache"
keywords = ["CRO", "blockchain", "crypto.com", "cosmos"]
readme = "README.md"
homepage = "https://github.com/crypto-com/pystarport"
repository = "https://github.com/crypto-com/pystarport"
packages = [{include = "pystarport"}]
include = ["pystarport/bin/*"]

[tool.poetry.dependencies]
python = "^3.9"
fire = "^0"
tomlkit = "^0"
jsonmerge = "^1.7"
python-dateutil = "^2.8"
durations = "^0"
supervisor = "^4.2"
docker = "^7.0"
bech32 = "^1.1"
multitail2 = "^1.5"
python-dotenv = "^1.0"
pyyaml-include = "^1.3"
jsonnet = "^0"
pyyaml = "^6.0"
grpcio = "^1.75.1"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
deepdiff = "^6.7"
flake8 = "^7"
flake8-black = "^0"
flake8-isort = "^6.1"

[tool.poetry.scripts]
pystarport = "pystarport.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
